Agregue el módulo utils y actualice Makefile para incluir utils.c
