+++ Makefile
@@ -1,6 +1,6 @@
 CC=gcc
 CFLAGS=-I./include -Wall -Wextra -O2
-SRC= main.c src/control_de_parametros.c
+SRC= main.c src/control_de_parametros.c src/utils.c
 TARGET=main
 
 all: $(TARGET)
